+++ unity/Assets/Scripts/Engine/Collision/CollisionManager.cs
@@ -9,7 +9,7 @@ public class CollisionManager : MonoSingleton<CollisionManager>
 {
     public bool HasCollision(ChunkInfo ci, int x, int y)
     {
-        Vector2 worldPos = GameManager.Chunk2World(ci, x, y);
+        Vector2 worldPos = WorldMap.Chunk2World(ci, x, y);
 
         Box2D worldBox = new Box2D(worldPos, 0.5f, 0.5f);
