+++ unity/Assets/Scripts/Game/Map/BiomeEnum.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+// Similar to biomes in Minecraft and Terraria: 
+// http://terraria.wikia.com/wiki/Environments
+// https://minecraft.gamepedia.com/Biome
+public enum EBiome
+{
+    Plain,
+    Ocean,
+    Forest,
+    Desert,
+    Mountain,
+    Snow,
+    Jungle,
+}
