+++ unity/Assets/Scripts/Game/Map/BiomeManager.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class BiomeManager : MonoBehaviour
+{
+
+
+    public BiomeManager()
+    {
+
+    }
+}
