+++ unity/Assets/Scripts/Game/Map/WorldMap.cs.meta
@@ -1,8 +1,11 @@
 fileFormatVersion: 2
-guid: 4e5355cbefd24094a97a752e6c632bdd
+guid: 30cc4dc913c02744997833470b36dcee
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
   icon: {instanceID: 0}
   userData: 
+  assetBundleName: 
+  assetBundleVariant:
